Extract more functions and name types consistently
